improve performance

+++ src/memory.h
@@ -3,7 +3,7 @@
 
 #include <string.h>
 
-static inline unsigned int count_for_strlcpy(char *point)
+static inline unsigned short count_for_strlcpy(char *point)
 {
   return (sizeof(char)*strlen(point)+sizeof(char));
 }
